requirements.txt
torch
pyyaml
h5py
numpy
